# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: P2P Exchange Ethiopia API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit-address:
    get:
      operationId: getDepositAddress
      tags: [wallet]
      summary: Get deposit address for network
      parameters:
        - name: network
          in: query
          required: true
          schema:
            type: string
            enum: [TRC20, ERC20]
      responses:
        "200":
          description: Deposit address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositAddress"
  /wallet/withdraw:
    post:
      operationId: requestWithdraw
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawInput"
      responses:
        "200":
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # Transactions
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile with stats
      responses:
        "200":
          description: Profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile (username)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [profile]
      summary: List user payment methods
      responses:
        "200":
          description: Payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      operationId: addPaymentMethod
      tags: [profile]
      summary: Add a payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodInput"
      responses:
        "201":
          description: Payment method added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /profile/payment-methods/{id}:
    delete:
      operationId: deletePaymentMethod
      tags: [profile]
      summary: Delete a payment method
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # KYC
...
components:
  schemas:
...
    DepositAddress:
      type: object
      properties:
        address:
          type: string
        network:
          type: string
        qrCode:
          type: string
        minDeposit:
          type: string
      required:
        - address
        - network
        - minDeposit
    WithdrawInput:
      type: object
      properties:
        address:
          type: string
        network:
          type: string
          enum: [TRC20, ERC20]
        amount:
          type: string
      required:
        - address
        - network
        - amount
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdraw, transfer]
        amount:
          type: string
        network:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed]
        txid:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        fee:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - amount
        - status
        - createdAt
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        kycStatus:
          type: string
          enum: [pending, verified, rejected, more_info_required, none]
        isMerchant:
          type: boolean
        emailVerified:
          type: boolean
        smsVerified:
          type: boolean
        addressVerified:
          type: boolean
        trades30d:
          type: integer
        completionRate30d:
          type: string
        avgReleaseTime:
          type: string
        avgPayTime:
          type: string
        totalTrades:
          type: integer
        positiveFeedbackPct:
          type: string
        negativeFeedbackPct:
          type: string
        feedbackCount:
          type: integer
        firstTradeAt:
          type: ["string", "null"]
        notificationSettings:
          type: object
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - kycStatus
        - isMerchant
        - emailVerified
        - smsVerified
        - addressVerified
        - trades30d
        - completionRate30d
        - avgReleaseTime
        - avgPayTime
        - totalTrades
        - positiveFeedbackPct
        - negativeFeedbackPct
        - feedbackCount
        - notificationSettings
        - createdAt
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
    PaymentMethod:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [CBE, Telebirr, Awash, Dashen, Abyssinia, HelloCash, MPesa]
        accountName:
          type: string
        accountNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - accountName
        - accountNumber
        - createdAt
    PaymentMethodInput:
      type: object
      properties:
        type:
          type: string
          enum: [CBE, Telebirr, Awash, Dashen, Abyssinia, HelloCash, MPesa]
        accountName:
          type: string
        accountNumber:
          type: string
      required:
        - type
        - accountName
        - accountNumber
...